WIP: fixed issue where login check was delayed
-In Callback.js, handleAuthentication() would complete after next line
where isAuthenticated was checked, causing the UI’s perception of the
login to be incorrect. This happened because handleAuthentication was
asynchronous (but using ‘async’ and ‘await’ keywords were not affecting
it, as expected).
-Solution was to add a callback function into handleAuthentication() so
that it would call a method that checks if the user was authenticated
only after handleAuthentication() was finished.
-This version uses a crude implementation as a fix; will clean up in
next commit.